make it clear we only have one storage object for the app, introduce signals for config
