+++ karmabot/__init__.py
@@ -4,17 +4,3 @@
 # This file is part of 'karmabot' and is distributed under the BSD license.
 # See LICENSE for more details.
 __import__('pkg_resources').declare_namespace(__name__)
-
-from .core import (
-    client,
-    commands,
-    facets,
-    ircutils,
-    thing,
-    register,
-    utils,
-)
-
-
-__all__ = ['client', 'commands', 'facets', 'ircutils',
-           'thing', 'register', 'utils']
