+++ karmabot/core/signal.py
@@ -1,0 +1,3 @@
+from blinker import signal
+
+post_connection = signal('post_connection')
